extract shared logic in pickbans dashboard into a shared composable
